Move clips manage page to template #292
