+++ web/static_files/clips/manage.html.jinja
@@ -1,0 +1,66 @@
+<html>
+<head>
+    <meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate" />
+</head>
+<body>
+    <script src="/clips/manage.js"></script>
+
+    <table border=1>
+    <tr><td colspan=10 align=center>
+        <div id="status">
+            Manage Clips.
+        </div>
+        </td>
+    </tr>
+        <tr>
+            <td>ID</td>
+            <td>Enabled</td>
+            <td>Name</td>
+            <td>Title</td>
+            <td>URL</td>
+            <td>Added By</td>
+            <td>Delete</td>
+        </tr>
+    {% for clip in result %}
+
+        <tr>
+            <td>{{clip.id}}</td>
+            <td>
+                <form action="" >
+                <input type="checkbox" name="enabled" onchange="updateClipEnable({{clip.id}}, this.checked)"
+                    {% if not user.admin %} disabled{% endif %}
+                    {% if clip.enabled %}checked{% endif %}>
+                </form>
+            </td>
+            <td>{{clip.name}}</td>
+            <td>{{clip.title}}</td>
+            <td><a href=https://clips.twitch.tv/{{clip.url}} target=_blank>{{clip.url}}</a></td>
+            <td>{{clip.added_by}}</td>
+            <td><form action="/clips/delete" method=POST>
+                <button name="clip_id" value={{clip.id}} onclick="return confirm('Do you really want to delete this clip?');">🗑️</button>
+                </form>
+            </td>
+        </tr>
+    {% endfor %}
+
+    </table>
+    <p>
+    <form method=POST action="/clips/add">
+    <table border=1>
+        <tr><td colspan=10 align=center>Insert new Clip</td></tr>
+        <tr>
+            <td>Enabled</td>
+            <td>Name</td>
+            <td>Title</td>
+            <td>URL</td>
+        </tr>
+        <tr>
+            <td><input type="checkbox" name="enabled"></td>
+            <td><input type="text" name="name"></td>
+            <td><input type="text" name="title"></td>
+            <td><input type="text" name="url"></td>
+            <td><input type="submit" value="Save"></td>
+        </tr>
+    </table>
+</body>
+</html>
